Tidy up sliding titles code
This includes very few visual changes, and mostly just reformats code to
be more Sass-y. The transition time is reduced to make it seem more
snappy, and a list-style with a dash is faked to make delineate articles
a little better.

We'll still have to set up infrastructure to use custom images (and
probably specify more than one for a specific category / specify custom
position attributes) and dump those into the template. Gonna be gross.