Uncomment old topics page CSS and add main class hack
We'll probably want to keep the old topics page in its current form
around somewhere, as its a nice compact way to view things, like a
sitemap. No telling where that will be yet, but may as well uncomment
the CSS.

Also, the container for the new topics page should really be full-width
on mobile, so we add a gross layout hack to enable that. This should
actually probably be the case for all pages, but its limited to the
topics page for now.