Fix navigation dropdown links
This links the dropdown to the correct sections on the new topics page.
I ran into the weirdest error while working on this: for example,
linking to /sexinfo/category#the_body is fine, whereas
/sexinfo/category/#the_body (note the trailing slash) totally breaks the
images with a ton of 404s. This can also be seen just by browsing to
/sexinfo/category and /sexinfo/category/. I think this is something
weird with base URIs and .htaccess rules, but I don't care to
investigate, and we've safe for now.