Installed jquery update module
This is required because the version of jquery packaged with Drupal is 1.4.2 while the latest is 1.11.2.  This means a lot of the code we wrote broke when we updated to Drupal 7.35 from 7.9.  This module will allow us to still use jQuery at the latest version even if Drupal chooses not to update with it.